inital work on non-ticket thread sync
