+++ tickets.py
@@ -9,7 +9,7 @@ import urllib.parse
 
 
 
-from model import SYNC_FIELD_NAME, TO_CC_FIELD_NAME, User, Message, NamedId, Team, Ticket, TicketNote, TicketsResult
+from model import TO_CC_FIELD_NAME, User, Message, NamedId, Team, Ticket, TicketNote, TicketsResult
 from session import RedmineSession, RedmineException
 import synctime
 
@@ -486,7 +486,7 @@ def main():
     #print(ticket_mgr.get(105, include_children=True).json_str())
     #print(json.dumps(ticket_mgr.load_custom_fields(), indent=4, default=vars))
 
-    #print(ticket_mgr.due())
+    print(ticket_mgr.due())
 
 # for testing the redmine
 if __name__ == '__main__':
